[REFACTORING] : Chat MVVM 완성
